+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Feature
 - [awesome] add TheMerge and l2beat
 - [contract] add bridge contracts from https://github.com/krzkaczor/l2beat
+- [hook] import url with jsonRPCProvider
 
 Refactor
 - update eth-scan eth/xdai/bsc/matic contract address